BB-186: Move login page into forms directory
